Moving sessions code into sub-package
Since the sessions code mostly stands on it's own, I wanted to break the
dependencies apart from it and move it into it's own package.
